build: do make lib/core/brython-runner.worker.js a bundle
Since this file is loaded as a raw string by
`lib/core/brython-runner.js`, calls like
`require("@babel/runtime/...")` will not be resolved by webpack and will
be kept in the resulting string. Because of that, the loaded string will
fail to run in a Web worker. In order to solve this, let us turn
`lib/core/brython-runner.worker.js` into a bundle, so that those
`require()` calls are already resolved during build time.